Refactor directory layout - enables client-only rebuilds, reduces rebuild time
