+++ client/ui/components/authenticated-navigation.js
