+++ client/ui/containers/documents-list.js
@@ -1,5 +1,5 @@
 import { composeWithTracker } from 'react-komposer';
-import { Documents } from '../../api/documents/documents.js';
+import { Documents } from '/imports/api/documents/documents.js';
 import { DocumentsList } from '../components/documents-list.js';
 import { Loading } from '../components/loading.js';
 import { Meteor } from 'meteor/meteor';
